+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MiniAppMessageBridge.kt
@@ -7,16 +7,16 @@ import androidx.annotation.VisibleForTesting
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.rakuten.tech.mobile.miniapp.CustomPermissionsNotImplementedException
+import com.rakuten.tech.mobile.miniapp.DevicePermissionsNotImplementedException
+import com.rakuten.tech.mobile.miniapp.PermissionsNotImplementedException
 import com.rakuten.tech.mobile.miniapp.ads.AdMobDisplayer
 import com.rakuten.tech.mobile.miniapp.ads.MiniAppAdDisplayer
 import com.rakuten.tech.mobile.miniapp.display.WebViewListener
 import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridgeDispatcher
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionResult
+import com.rakuten.tech.mobile.miniapp.permission.*
 import com.rakuten.tech.mobile.miniapp.permission.CustomPermissionBridgeDispatcher
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionCache
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionResult
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppDevicePermissionResult
 import com.rakuten.tech.mobile.miniapp.permission.ui.MiniAppCustomPermissionWindow
 
 @Suppress("TooGenericExceptionCaught", "ComplexMethod", "LargeClass", "TooManyFunctions", "LongMethod")
@@ -75,10 +75,21 @@ abstract class MiniAppMessageBridge {
     abstract fun getUniqueId(): String
 
     /** Post permission request from external. **/
-    abstract fun requestPermission(
+    @Deprecated("Use requestDevicePermission instead")
+    open fun requestPermission(
         miniAppPermissionType: MiniAppPermissionType,
         callback: (isGranted: Boolean) -> Unit
-    )
+    ) {
+        throw PermissionsNotImplementedException()
+    }
+
+    /** Post device permission request from external. **/
+    open fun requestDevicePermission(
+        miniAppPermissionType: MiniAppDevicePermissionType,
+        callback: (isGranted: Boolean) -> Unit
+    ) {
+        throw DevicePermissionsNotImplementedException()
+    }
 
     /**
      * Post custom permissions request.
@@ -124,6 +135,7 @@ abstract class MiniAppMessageBridge {
         when (callbackObj.action) {
             ActionType.GET_UNIQUE_ID.action -> onGetUniqueId(callbackObj)
             ActionType.REQUEST_PERMISSION.action -> onRequestPermission(callbackObj)
+            ActionType.REQUEST_DEVICE_PERMISSION.action -> onRequestDevicePermission(callbackObj)
             ActionType.REQUEST_CUSTOM_PERMISSIONS.action -> onRequestCustomPermissions(jsonStr)
             ActionType.SHARE_INFO.action -> onShareContent(callbackObj.id, jsonStr)
             ActionType.LOAD_AD.action -> adBridgeDispatcher.onLoadAd(callbackObj.id, jsonStr)
@@ -159,14 +171,32 @@ abstract class MiniAppMessageBridge {
 
     private fun onRequestPermission(callbackObj: CallbackObj) {
         try {
-            val permissionParam = Gson().fromJson<Permission>(
+            val permissionParam = Gson().fromJson<DevicePermission>(
                 callbackObj.param.toString(),
-                object : TypeToken<Permission>() {}.type
+                object : TypeToken<DevicePermission>() {}.type
             )
 
             requestPermission(
                 MiniAppPermissionType.getValue(permissionParam.permission)
-            ) { isGranted -> onRequestPermissionsResult(
+            ) { isGranted -> onRequestDevicePermissionsResult(
+                callbackId = callbackObj.id,
+                isGranted = isGranted
+            ) }
+        } catch (e: Exception) {
+            bridgeExecutor.postError(callbackObj.id, "${ErrorBridgeMessage.ERR_REQ_PERMISSION} ${e.message}")
+        }
+    }
+
+    private fun onRequestDevicePermission(callbackObj: CallbackObj) {
+        try {
+            val permissionParam = Gson().fromJson<DevicePermission>(
+                callbackObj.param.toString(),
+                object : TypeToken<DevicePermission>() {}.type
+            )
+
+            requestDevicePermission(
+                MiniAppDevicePermissionType.getValue(permissionParam.permission)
+            ) { isGranted -> onRequestDevicePermissionsResult(
                 callbackId = callbackObj.id,
                 isGranted = isGranted
             ) }
@@ -222,11 +252,11 @@ abstract class MiniAppMessageBridge {
 
     @VisibleForTesting
     /** Inform the permission request result to MiniApp. **/
-    internal fun onRequestPermissionsResult(callbackId: String, isGranted: Boolean) {
+    internal fun onRequestDevicePermissionsResult(callbackId: String, isGranted: Boolean) {
         if (isGranted)
-            bridgeExecutor.postValue(callbackId, MiniAppPermissionResult.getValue(isGranted).type)
+            bridgeExecutor.postValue(callbackId, MiniAppDevicePermissionResult.getValue(isGranted).type)
         else
-            bridgeExecutor.postError(callbackId, MiniAppPermissionResult.getValue(isGranted).type)
+            bridgeExecutor.postError(callbackId, MiniAppDevicePermissionResult.getValue(isGranted).type)
     }
 
     internal fun onWebViewDetach() {
@@ -246,6 +276,12 @@ internal object ErrorBridgeMessage {
     const val ERR_UNIQUE_ID = "Cannot get unique id:"
     const val ERR_REQ_PERMISSION = "Cannot request permission:"
     const val ERR_REQ_CUSTOM_PERMISSION = "Cannot request custom permissions:"
+    const val NO_IMPLEMENT_PERMISSION =
+        "The `MiniAppMessageBridge.requestPermission`" +
+                " method has not been implemented by the Host App."
+    const val NO_IMPLEMENT_DEVICE_PERMISSION =
+        "The `MiniAppMessageBridge.requestDevicePermission`" +
+                " method has not been implemented by the Host App."
     const val NO_IMPLEMENT_CUSTOM_PERMISSION =
         "The `MiniAppMessageBridge.requestCustomPermissions`" +
                 " method has not been implemented by the Host App."