+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/Type.kt
@@ -2,7 +2,8 @@ package com.rakuten.tech.mobile.miniapp.js
 
 internal enum class ActionType(val action: String) {
     GET_UNIQUE_ID("getUniqueId"),
-    REQUEST_PERMISSION("requestPermission"),
+    @Deprecated("Use REQUEST_DEVICE_PERMISSION instead") REQUEST_PERMISSION("requestPermission"),
+    REQUEST_DEVICE_PERMISSION("requestDevicePermission"),
     REQUEST_CUSTOM_PERMISSIONS("requestCustomPermissions"),
     SHARE_INFO("shareInfo"),
     LOAD_AD("loadAd"),