+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/helper/AppPermission.kt
@@ -1,6 +1,7 @@
 package com.rakuten.tech.mobile.testapp.helper
 
 import android.Manifest
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppDevicePermissionType
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
 
 object AppPermission {
@@ -18,4 +19,15 @@ object AppPermission {
         MiniAppPermissionType.LOCATION -> arrayOf(Manifest.permission.ACCESS_FINE_LOCATION)
         else -> emptyArray()
     }
+
+    fun getDeviceRequestCode(permissionType: MiniAppDevicePermissionType) = when (permissionType) {
+        MiniAppDevicePermissionType.LOCATION -> ReqCode.LOCATION
+        else -> 0
+    }
+
+    fun getDevicePermissionRequest(permissionType: MiniAppDevicePermissionType) =
+        when (permissionType) {
+            MiniAppDevicePermissionType.LOCATION -> arrayOf(Manifest.permission.ACCESS_FINE_LOCATION)
+            else -> emptyArray()
+        }
 }