+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayActivity.kt
@@ -24,6 +24,7 @@ import com.rakuten.tech.mobile.miniapp.js.userinfo.TokenData
 import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridgeDispatcher
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
 import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppDevicePermissionType
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.MiniAppDisplayActivityBinding
 import com.rakuten.tech.mobile.testapp.helper.AppPermission
@@ -153,6 +154,18 @@ class MiniAppDisplayActivity : BaseActivity() {
                     this@MiniAppDisplayActivity,
                     AppPermission.getPermissionRequest(miniAppPermissionType),
                     AppPermission.getRequestCode(miniAppPermissionType)
+                )
+            }
+
+            override fun requestDevicePermission(
+                miniAppPermissionType: MiniAppDevicePermissionType,
+                callback: (isGranted: Boolean) -> Unit
+            ) {
+                miniappPermissionCallback = callback
+                ActivityCompat.requestPermissions(
+                    this@MiniAppDisplayActivity,
+                    AppPermission.getDevicePermissionRequest(miniAppPermissionType),
+                    AppPermission.getDeviceRequestCode(miniAppPermissionType)
                 )
             }
         }